Move pkg/api/v1alpha1/ to pkg/apis/gadget/v1alpha1
That's the path needed to make it work with client-gen

TODO: should the controller codegen be updated?
