+++ pkg/gadgets/helpers.go
@@ -17,7 +17,7 @@ package gadgets
 import (
 	"fmt"
 
-	gadgetv1alpha1 "github.com/kinvolk/inspektor-gadget/pkg/api/v1alpha1"
+	gadgetv1alpha1 "github.com/kinvolk/inspektor-gadget/pkg/apis/gadget/v1alpha1"
 	pb "github.com/kinvolk/inspektor-gadget/pkg/gadgettracermanager/api"
 	"k8s.io/apimachinery/pkg/types"
 )
